[WIP] Refactor to remove  one `unsafe` and some index-juggling.
In some code paths, `interpreter::resolve` keeps track of a current
`subrange` of `Value`s. Since this subrange is stored as a pair of
`usize`s, Rust cannot keep track of the fact that the start and end
are continuously valid indices. This entails that the eventual
lookup of the range requires either `unsafe` code, or runtime bounds
checks.

This commit achieves the same behavior by modeling the subrange as
a `&[Value]` that serves as a view on the input `Value`.

It is a [WIP], since the error messages are temporarily broken
by the changes.