[WIP] Removes a use of `unsafe` which lies about mutability.
Low-key lying about mutability to the compiler ain't the best idea.
UB abound. This commit handles one case where this happens, the
painful way: since the only reason that the `&TremorAggrFnWrapper`
was magiced into a mutable reference is to be able to mutate its
`cache` field, the `unsafe` `transmute` can be removed if we're
up-front to the compiler that `cache` is actually internally mutable,
using a `RefCell`. In hindsight, it would probably make more sense to
make the `invocable` field a `RefCell<TremorAggrFnWrapper>` instead,
since then we'd only need one runtime mutable borrow per invocation,
instead of many internally as this commit achieves. Live and Learn™.