[WIP] Attempt #2 for subrange-as-slice in interpreter::resolve.
The error messages on out-of-bounds access are correct as they are.
The returned error message is verified by existing tests:
- for `Segment::Idx`: tests `script_runtime_error/arr_bad_idx` and
  `script_runtime_error/subslice_and_idx_out_of_bounds`,
- for `Segment::Element` where `current` is a `Value::Array`: test
  `script_runtime_error/arr_bad_idx2`, and
- for `Segment::Range`: test
  `script_runtime_error/subslice_out_of_bounds`.

However, some other boundary conditions were not yet handled in the
existing code before this commit:
a. if the start index of a sub-range is a negative integer it would
   give the confusing error that it "expected an integer, but got an
   integer" as the conversion to `u64` would fail, but be handled by
   `error_need_int`.
b. if the end index was smaller than the start index, this would first
   be checked by the construction of a Rust-level range, which would
   panic the runtime.

This commit 'handles' those cases, but poorly. Two new error-message
kinds need to be added.

Awaiting feedback.